# rest/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Chainspace REST API endpoints",
        "title": "Chainspace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/cs-coin/wallet": {
            "post": {
                "description": "Create a new Wallet from address generated by the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cs-coin-methods"
                ],
                "summary": "Create a new Wallet from address generated by the user",
                "operationId": "createWallet",
                "parameters": [
                    {
                        "description": "wallet",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.CreateWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/cs-coin/wallet/add-funds": {
            "post": {
                "description": "Add new funds to user wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cs-coin-methods"
                ],
                "summary": "Add new funds to a user wallet",
                "operationId": "addFunds",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.AddFundsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/cs-coin/wallet/init/{addr}": {
            "get": {
                "description": "Generate the seed object for a new wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cs-coin-init"
                ],
                "summary": "Generate the seed object for a new wallet",
                "operationId": "init",
                "parameters": [
                    {
                        "type": "string",
                        "description": "addr",
                        "name": "addr",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.InitResponse"
                        }
                    }
                }
            }
        },
        "/api/cs-coin/wallet/transfer-funds": {
            "post": {
                "description": "Transfer funds from a wallet to another",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cs-coin-methods"
                ],
                "summary": "Transfer funds from a wallet to another",
                "operationId": "transferFunds",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.TransferFundsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AddFundsRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "mappings": {
                    "type": "object"
                },
                "signature": {
                    "type": "string"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
...
        "api.CreateWalletRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "initObject": {
                    "type": "string"
                },
                "mappings": {
                    "type": "object"
                },
                "pubKey": {
                    "type": "string"
                }
            }
        },
...
        "api.InitResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "initObject": {
                    "type": "string"
                }
            }
        },
...
        "api.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "transaction": {
                    "type": "string"
                }
            }
        },
...
        "api.TransferFundsRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "fromWallet": {
                    "type": "string"
                },
                "mappings": {
                    "type": "object"
                },
                "signature": {
                    "type": "string"
                },
                "toWallet": {
                    "type": "string"
                }
            }
        },
...
    }
}